Move all forms tables to FormTables folder. Remove old versions. Add Awaiting Approval table and container.
